refactor shard/mod.rs
