+++ lib/collection/src/lib.rs
@@ -9,7 +9,7 @@ pub mod hash_ring;
 pub mod operations;
 pub mod optimizers_builder;
 pub mod save_on_disk;
-pub mod shard;
+pub mod shards;
 pub mod telemetry;
 mod update_handler;
 pub mod wal;
