--- lib/collection/src/shard/mod.rs
@@ -1,235 +1,0 @@
-pub mod collection_shard_distribution;
-mod conversions;
-pub mod forward_proxy_shard;
-pub mod local_shard;
-pub mod local_shard_operations;
-pub mod proxy_shard;
-pub mod remote_shard;
-#[allow(dead_code)]
-pub mod replica_set;
-pub mod shard_config;
-pub mod shard_holder;
-pub mod shard_versioning;
-pub mod transfer;
-
-use std::collections::HashMap;
-use std::path::{Path, PathBuf};
-use std::sync::Arc;
-
-use api::grpc::transport_channel_pool::TransportChannelPool;
-use async_trait::async_trait;
-use schemars::JsonSchema;
-use segment::types::{
-    ExtendedPointId, Filter, ScoredPoint, WithPayload, WithPayloadInterface, WithVector,
-};
-use serde::{Deserialize, Serialize};
-use tokio::runtime::Handle;
-use tonic::transport::Uri;
-
-use self::replica_set::ReplicaSet;
-use crate::operations::types::{
-    CollectionError, CollectionInfo, CollectionResult, CountRequest, CountResult, PointRequest,
-    Record, SearchRequestBatch, UpdateResult,
-};
-use crate::operations::CollectionUpdateOperations;
-use crate::shard::forward_proxy_shard::ForwardProxyShard;
-use crate::shard::local_shard::LocalShard;
-use crate::shard::proxy_shard::ProxyShard;
-use crate::shard::remote_shard::RemoteShard;
-use crate::shard::shard_versioning::suggest_next_version_path;
-use crate::telemetry::ShardTelemetry;
-
-pub type ShardId = u32;
-
-/// Shard
-///
-/// Contains a part of the collection's points
-///
-#[allow(clippy::large_enum_variant)]
-pub enum Shard {
-    Local(LocalShard),
-    Remote(RemoteShard),
-    Proxy(ProxyShard),
-    ForwardProxy(ForwardProxyShard),
-    ReplicaSet(ReplicaSet),
-}
-
-impl Shard {
-    pub fn variant_name(&self) -> &str {
-        match self {
-            Shard::Local(_) => "local shard",
-            Shard::Remote(_) => "remote shard",
-            Shard::Proxy(_) => "proxy shard",
-            Shard::ForwardProxy(_) => "forward proxy shard",
-            Shard::ReplicaSet(_) => "replica set",
-        }
-    }
-
-    pub fn get(&self) -> &(dyn ShardOperation + Sync + Send + '_) {
-        match self {
-            Shard::Local(local_shard) => local_shard,
-            Shard::Remote(remote_shard) => remote_shard,
-            Shard::Proxy(proxy_shard) => proxy_shard,
-            Shard::ForwardProxy(proxy_shard) => proxy_shard,
-            Shard::ReplicaSet(replica_set) => replica_set,
-        }
-    }
-
-    pub async fn before_drop(&mut self) {
-        match self {
-            Shard::Local(local_shard) => local_shard.before_drop().await,
-            Shard::Remote(_) => (),
-            Shard::Proxy(proxy_shard) => proxy_shard.before_drop().await,
-            Shard::ForwardProxy(proxy_shard) => proxy_shard.before_drop().await,
-            Shard::ReplicaSet(shard) => shard.before_drop().await,
-        }
-    }
-
-    pub fn peer_ids(&self, this_peer_id: PeerId) -> Vec<PeerId> {
-        match self {
-            Shard::Local(_) => vec![this_peer_id],
-            Shard::Remote(remote) => vec![remote.peer_id],
-            Shard::Proxy(_) => vec![this_peer_id],
-            Shard::ForwardProxy(_) => vec![this_peer_id],
-            Shard::ReplicaSet(replicas) => replicas.peer_ids(),
-        }
-    }
-
-    pub fn get_telemetry_data(&self) -> ShardTelemetry {
-        match self {
-            Shard::Local(local_shard) => local_shard.get_telemetry_data(),
-            Shard::Remote(remote_shard) => remote_shard.get_telemetry_data(),
-            Shard::Proxy(proxy_shard) => proxy_shard.get_telemetry_data(),
-            Shard::ForwardProxy(proxy_shard) => proxy_shard.get_telemetry_data(),
-            Shard::ReplicaSet(replica_set) => replica_set.get_telemetry_data(),
-        }
-    }
-
-    pub async fn create_snapshot(&self, target_path: &Path) -> CollectionResult<()> {
-        match self {
-            Shard::Local(local_shard) => local_shard.create_snapshot(target_path).await,
-            Shard::Remote(remote_shard) => remote_shard.create_snapshot(target_path).await,
-            Shard::Proxy(proxy_shard) => proxy_shard.create_snapshot(target_path).await,
-            Shard::ForwardProxy(proxy_shard) => proxy_shard.create_snapshot(target_path).await,
-            Shard::ReplicaSet(replica_set) => replica_set.create_snapshot(target_path).await,
-        }
-    }
-
-    pub async fn on_optimizer_config_update(&self) -> CollectionResult<()> {
-        match self {
-            Shard::Local(local_shard) => local_shard.on_optimizer_config_update().await,
-            Shard::Remote(_) => Ok(()),
-            Shard::Proxy(proxy_shard) => proxy_shard.on_optimizer_config_update().await,
-            Shard::ForwardProxy(proxy_shard) => proxy_shard.on_optimizer_config_update().await,
-            Shard::ReplicaSet(shard) => shard.on_optimizer_config_update().await,
-        }
-    }
-}
-
-#[async_trait]
-pub trait ShardOperation {
-    async fn update(
-        &self,
-        operation: CollectionUpdateOperations,
-        wait: bool,
-    ) -> CollectionResult<UpdateResult>;
-
-    #[allow(clippy::too_many_arguments)]
-    async fn scroll_by(
-        &self,
-        offset: Option<ExtendedPointId>,
-        limit: usize,
-        with_payload_interface: &WithPayloadInterface,
-        with_vector: &WithVector,
-        filter: Option<&Filter>,
-    ) -> CollectionResult<Vec<Record>>;
-
-    async fn info(&self) -> CollectionResult<CollectionInfo>;
-
-    async fn search(
-        &self,
-        request: Arc<SearchRequestBatch>,
-        search_runtime_handle: &Handle,
-    ) -> CollectionResult<Vec<Vec<ScoredPoint>>>;
-
-    async fn count(&self, request: Arc<CountRequest>) -> CollectionResult<CountResult>;
-
-    async fn retrieve(
-        &self,
-        request: Arc<PointRequest>,
-        with_payload: &WithPayload,
-        with_vector: &WithVector,
-    ) -> CollectionResult<Vec<Record>>;
-}
-
-pub const HASH_RING_SHARD_SCALE: u32 = 100;
-
-pub type CollectionId = String;
-
-pub type ShardVersion = usize;
-
-pub type PeerId = u64;
-
-#[derive(Clone)]
-pub struct ChannelService {
-    pub id_to_address: Arc<parking_lot::RwLock<HashMap<PeerId, Uri>>>,
-    pub channel_pool: Arc<TransportChannelPool>,
-}
-
-impl ChannelService {
-    pub fn new(
-        id_to_address: Arc<parking_lot::RwLock<HashMap<PeerId, Uri>>>,
-        channel_pool: Arc<TransportChannelPool>,
-    ) -> Self {
-        Self {
-            id_to_address,
-            channel_pool,
-        }
-    }
-
-    pub async fn remove_peer(&self, peer_id: PeerId) {
-        let removed = self.id_to_address.write().remove(&peer_id);
-        if let Some(uri) = removed {
-            self.channel_pool.drop_pool(&uri).await;
-        }
-    }
-}
-
-impl Default for ChannelService {
-    fn default() -> Self {
-        Self {
-            id_to_address: Arc::new(Default::default()),
-            channel_pool: Arc::new(Default::default()),
-        }
-    }
-}
-
-#[derive(Debug, Clone, Hash, PartialEq, Eq, Serialize, Deserialize, JsonSchema)]
-pub struct ShardTransfer {
-    pub shard_id: ShardId,
-    pub from: PeerId,
-    pub to: PeerId,
-    /// If this flag is true, the is a replication related transfer of shard from 1 peer to another
-    /// Shard on original peer will not be deleted in this case
-    pub sync: bool,
-}
-
-pub async fn create_shard_dir(
-    collection_path: &Path,
-    shard_id: ShardId,
-) -> CollectionResult<PathBuf> {
-    loop {
-        let shard_path = suggest_next_version_path(collection_path, shard_id).await?;
-
-        match tokio::fs::create_dir(&shard_path).await {
-            Ok(_) => return Ok(shard_path),
-            Err(e) => {
-                if e.kind() == std::io::ErrorKind::AlreadyExists {
-                    continue;
-                } else {
-                    return Err(CollectionError::from(e));
-                }
-            }
-        }
-    }
-}
