+++ lib/collection/src/shards/local_shard.rs
@@ -25,8 +25,9 @@ use crate::config::CollectionConfig;
 use crate::operations::types::{CollectionError, CollectionResult};
 use crate::operations::CollectionUpdateOperations;
 use crate::optimizers_builder::build_optimizers;
-use crate::shard::shard_config::{ShardConfig, SHARD_CONFIG_FILE};
-use crate::shard::{CollectionId, ShardId};
+use crate::shards::shard::ShardId;
+use crate::shards::shard_config::{ShardConfig, SHARD_CONFIG_FILE};
+use crate::shards::CollectionId;
 use crate::telemetry::{LocalShardTelemetry, ShardTelemetry};
 use crate::update_handler::{Optimizer, UpdateHandler, UpdateSignal, UPDATE_QUEUE_SIZE};
 use crate::wal::SerdeWal;
