+++ lib/collection/src/shards/remote_shard.rs
@@ -26,14 +26,17 @@ use crate::operations::types::{
     Record, SearchRequest, SearchRequestBatch, UpdateResult,
 };
 use crate::operations::{CollectionUpdateOperations, FieldIndexOperations};
-use crate::shard::conversions::{
+use crate::shards::channel_service::ChannelService;
+use crate::shards::conversions::{
     internal_clear_payload, internal_clear_payload_by_filter, internal_create_index,
     internal_delete_index, internal_delete_payload, internal_delete_points,
     internal_delete_points_by_filter, internal_set_payload, internal_sync_points,
     internal_upsert_points,
 };
-use crate::shard::shard_config::ShardConfig;
-use crate::shard::{ChannelService, CollectionId, PeerId, ShardId, ShardOperation};
+use crate::shards::shard::{PeerId, ShardId};
+use crate::shards::shard_config::ShardConfig;
+use crate::shards::shard_trait::ShardOperation;
+use crate::shards::CollectionId;
 use crate::telemetry::{RemoteShardTelemetry, ShardTelemetry};
 
 /// RemoteShard
