+++ lib/collection/src/shards/shard_config.rs
@@ -4,7 +4,7 @@ use segment::common::file_operations::{atomic_save_json, read_json};
 use serde::{Deserialize, Serialize};
 
 use crate::operations::types::CollectionResult;
-use crate::shard::PeerId;
+use crate::shards::shard::PeerId;
 
 pub const SHARD_CONFIG_FILE: &str = "shard_config.json";
