+++ lib/collection/src/shards/shard_trait.rs
@@ -1,0 +1,49 @@
+use std::sync::Arc;
+
+use async_trait::async_trait;
+use segment::types::{
+    ExtendedPointId, Filter, ScoredPoint, WithPayload, WithPayloadInterface, WithVector,
+};
+use tokio::runtime::Handle;
+
+use crate::operations::types::{
+    CollectionInfo, CollectionResult, CountRequest, CountResult, PointRequest, Record,
+    SearchRequestBatch, UpdateResult,
+};
+use crate::operations::CollectionUpdateOperations;
+
+#[async_trait]
+pub trait ShardOperation {
+    async fn update(
+        &self,
+        operation: CollectionUpdateOperations,
+        wait: bool,
+    ) -> CollectionResult<UpdateResult>;
+
+    #[allow(clippy::too_many_arguments)]
+    async fn scroll_by(
+        &self,
+        offset: Option<ExtendedPointId>,
+        limit: usize,
+        with_payload_interface: &WithPayloadInterface,
+        with_vector: &WithVector,
+        filter: Option<&Filter>,
+    ) -> CollectionResult<Vec<Record>>;
+
+    async fn info(&self) -> CollectionResult<CollectionInfo>;
+
+    async fn search(
+        &self,
+        request: Arc<SearchRequestBatch>,
+        search_runtime_handle: &Handle,
+    ) -> CollectionResult<Vec<Vec<ScoredPoint>>>;
+
+    async fn count(&self, request: Arc<CountRequest>) -> CollectionResult<CountResult>;
+
+    async fn retrieve(
+        &self,
+        request: Arc<PointRequest>,
+        with_payload: &WithPayload,
+        with_vector: &WithVector,
+    ) -> CollectionResult<Vec<Record>>;
+}
