+++ lib/storage/src/content_manager/alias_mapping.rs
@@ -3,7 +3,7 @@ use std::fs;
 use std::io::Write;
 use std::path::{Path, PathBuf};
 
-use collection::shard::CollectionId;
+use collection::shards::CollectionId;
 use segment::common::file_operations::{atomic_save_json, read_json};
 use serde::{Deserialize, Serialize};
