+++ lib/storage/src/content_manager/mod.rs
@@ -1,4 +1,4 @@
-use collection::shard::PeerId;
+use collection::shards::shard::PeerId;
 
 use self::collection_meta_ops::CollectionMetaOperations;
 use self::consensus_state::CollectionsSnapshot;
@@ -16,7 +16,8 @@ pub mod snapshots;
 pub mod toc;
 
 pub mod consensus_ops {
-    use collection::shard::{CollectionId, PeerId, ShardTransfer};
+    use collection::shards::shard::PeerId;
+    use collection::shards::{CollectionId, ShardTransfer};
     use raft::eraftpb::Entry as RaftEntry;
     use serde::{Deserialize, Serialize};
