+++ src/common/collections.rs
@@ -7,7 +7,8 @@ use collection::operations::cluster_ops::{
 };
 use collection::operations::snapshot_ops::SnapshotDescription;
 use collection::operations::types::{CollectionClusterInfo, CollectionInfo};
-use collection::shard::{replica_set, ShardId, ShardTransfer};
+use collection::shards::shard::ShardId;
+use collection::shards::{replica_set, ShardTransfer};
 use itertools::Itertools;
 use storage::content_manager::collection_meta_ops::ShardTransferOperations::{Abort, Start};
 use storage::content_manager::collection_meta_ops::{
